Move UDP test to single file, similar to `tcp_sample_application.rs`
