+++ crates/test-programs/tests/wasi-sockets.rs
@@ -92,13 +92,8 @@ async fn tcp_sockopts() {
 }
 
 #[test_log::test(tokio::test(flavor = "multi_thread"))]
-async fn udp_v4() {
-    run("udp_v4").await.unwrap();
-}
-
-#[test_log::test(tokio::test(flavor = "multi_thread"))]
-async fn udp_v6() {
-    run("udp_v6").await.unwrap();
+async fn udp_sample_application() {
+    run("udp_sample_application").await.unwrap();
 }
 
 #[test_log::test(tokio::test(flavor = "multi_thread"))]
